+++ libs/ymir-core/src/ymir/hw/vdp/vdp.cpp
@@ -90,6 +90,9 @@ VDP::~VDP() {
         if (m_VDPRenderThread.joinable()) {
             m_VDPRenderThread.join();
         }
+        if (m_VDPDeinterlaceRenderThread.joinable()) {
+            m_VDPDeinterlaceRenderThread.join();
+        }
     }
 }
 
@@ -119,6 +122,10 @@ void VDP::Reset(bool hard) {
     m_spriteLayerState[1].Reset();
     for (auto &state : m_normBGLayerStates) {
         state.Reset();
+    }
+    for (auto &state : m_charFetchers) {
+        state[0].Reset();
+        state[1].Reset();
     }
     for (auto &state : m_rotParamStates) {
         state.Reset();
@@ -502,11 +509,15 @@ void VDP::EnableThreadedVDP(bool enable) {
         m_VDPRenderContext.EnqueueEvent(VDPRenderEvent::UpdateEffectiveRenderingFlags());
         m_VDPRenderContext.EnqueueEvent(VDPRenderEvent::PostLoadStateSync());
         m_VDPRenderThread = std::thread{[&] { VDPRenderThread(); }};
+        m_VDPDeinterlaceRenderThread = std::thread{[&] { VDPDeinterlaceRenderThread(); }};
         m_VDPRenderContext.postLoadSyncSignal.Wait(true);
     } else {
         m_VDPRenderContext.EnqueueEvent(VDPRenderEvent::Shutdown());
         if (m_VDPRenderThread.joinable()) {
             m_VDPRenderThread.join();
+        }
+        if (m_VDPDeinterlaceRenderThread.joinable()) {
+            m_VDPDeinterlaceRenderThread.join();
         }
 
         VDPRenderEvent dummy{};
@@ -859,9 +870,14 @@ void VDP::BeginHPhaseActiveDisplay() {
 
             m_VDPRenderContext.EnqueueEvent(VDPRenderEvent::VDP2DrawLine(m_state.regs2.VCNT));
         } else {
+            const bool interlaced = m_state.regs2.TVMD.IsInterlaced();
             const uint32 y = m_state.regs2.VCNT;
             VDP2PrepareLine(y);
-            (this->*m_fnVDP2DrawLine)(y);
+            (this->*m_fnVDP2DrawLine)(y, false);
+            if (m_deinterlaceRender && interlaced) {
+                (this->*m_fnVDP2DrawLine)(y, true);
+            }
+            VDP2FinishLine(y);
         }
     }
 }
@@ -1067,10 +1083,22 @@ void VDP::VDPRenderThread() {
 
             case EvtType::VDP2BeginFrame: VDP2CalcAccessPatterns(rctx.vdp2.regs); break;
             case EvtType::VDP2UpdateEnabledBGs: VDP2UpdateEnabledBGs(); break;
-            case EvtType::VDP2DrawLine:
+            case EvtType::VDP2DrawLine: //
+            {
+                const bool deinterlaceRender = m_deinterlaceRender;
+                const bool interlaced = rctx.vdp2.regs.TVMD.IsInterlaced();
                 VDP2PrepareLine(event.drawLine.vcnt);
-                (this->*m_fnVDP2DrawLine)(event.drawLine.vcnt);
+                if (deinterlaceRender && interlaced) {
+                    rctx.deinterlaceY = event.drawLine.vcnt;
+                    rctx.deinterlaceRenderBeginSignal.Set();
+                }
+                (this->*m_fnVDP2DrawLine)(event.drawLine.vcnt, false);
+                if (deinterlaceRender && interlaced) {
+                    rctx.deinterlaceRenderEndSignal.Wait(true);
+                }
+                VDP2FinishLine(event.drawLine.vcnt);
                 break;
+            }
             case EvtType::VDP2EndFrame: rctx.renderFinishedSignal.Set(); break;
 
             case EvtType::VDP1VRAMWriteByte: rctx.vdp1.VRAM[event.write.address] = event.write.value; break;
@@ -1160,9 +1188,32 @@ void VDP::VDPRenderThread() {
 
             case EvtType::UpdateEffectiveRenderingFlags: UpdateEffectiveRenderingFlags(); break;
 
-            case EvtType::Shutdown: running = false; break;
+            case EvtType::Shutdown:
+                rctx.deinterlaceShutdown = true;
+                rctx.deinterlaceRenderBeginSignal.Set();
+                rctx.deinterlaceRenderEndSignal.Wait(true);
+                running = false;
+                break;
             }
         }
+    }
+}
+
+void VDP::VDPDeinterlaceRenderThread() {
+    util::SetCurrentThreadName("VDP deinterlace render thread");
+
+    auto &rctx = m_VDPRenderContext;
+
+    while (true) {
+        rctx.deinterlaceRenderBeginSignal.Wait(true);
+        if (rctx.deinterlaceShutdown) {
+            rctx.deinterlaceShutdown = false;
+            rctx.deinterlaceRenderEndSignal.Set();
+            return;
+        }
+
+        (this->*m_fnVDP2DrawLine)(rctx.deinterlaceY, true);
+        rctx.deinterlaceRenderEndSignal.Set();
     }
 }
 
@@ -2535,6 +2586,19 @@ void VDP::VDP2UpdateEnabledBGs() {
     }
 }
 
+FORCE_INLINE void VDP::VDP2UpdateLineScreenScrollParams(uint32 y) {
+    const VDP2Regs &regs2 = VDP2GetRegs();
+
+    for (uint32 i = 0; i < 2; ++i) {
+        if (!m_layerEnabled[i + 2]) {
+            continue;
+        }
+        const BGParams &bgParams = regs2.bgParams[i + 1];
+        NormBGLayerState &bgState = m_normBGLayerStates[i];
+        VDP2UpdateLineScreenScroll(y, bgParams, bgState);
+    }
+}
+
 FORCE_INLINE void VDP::VDP2UpdateLineScreenScroll(uint32 y, const BGParams &bgParams, NormBGLayerState &bgState) {
     if ((y & ((1u << bgParams.lineScrollInterval) - 1)) != 0) {
         return;
@@ -3150,10 +3214,40 @@ FORCE_INLINE void VDP::VDP2PrepareLine(uint32 y) {
 
     // Draw line color and back screen layers
     VDP2DrawLineColorAndBackScreens(y);
+
+    VDP2UpdateLineScreenScrollParams(y);
+}
+
+FORCE_INLINE void VDP::VDP2FinishLine(uint32 y) {
+    const VDP2Regs &regs2 = VDP2GetRegs();
+    const bool doubleDensity = regs2.TVMD.LSMDn == InterlaceMode::DoubleDensity;
+
+    // Update NBG coordinates
+    for (uint32 i = 0; i < 4; ++i) {
+        if (!m_layerEnabled[i + 2]) {
+            continue;
+        }
+        const BGParams &bgParams = regs2.bgParams[i + 1];
+        NormBGLayerState &bgState = m_normBGLayerStates[i];
+        bgState.fracScrollY += bgParams.scrollIncV;
+        // Update the vertical scroll coordinate twice in double-density interlaced mode.
+        // If deinterlacing, the second increment is done after rendering the alternate scanline.
+        if (doubleDensity) {
+            bgState.fracScrollY += bgParams.scrollIncV;
+        }
+
+        // Increment mosaic counter
+        if (bgParams.mosaicEnable) {
+            ++bgState.mosaicCounterY;
+            if (bgState.mosaicCounterY >= regs2.mosaicV) {
+                bgState.mosaicCounterY = 0;
+            }
+        }
+    }
 }
 
 template <bool deinterlace, bool transparentMeshes>
-void VDP::VDP2DrawLine(uint32 y) {
+void VDP::VDP2DrawLine(uint32 y, bool altField) {
     devlog::trace<grp::vdp2_render>("Drawing line {}", y);
 
     const VDP1Regs &regs1 = VDP1GetRegs();
@@ -3187,98 +3281,36 @@ void VDP::VDP2DrawLine(uint32 y) {
     const bool doubleDensity = regs2.TVMD.LSMDn == InterlaceMode::DoubleDensity;
 
     // Precalculate window state
-    VDP2CalcWindows<deinterlace, false>(y);
+    if (altField) {
+        VDP2CalcWindows<deinterlace, true>(y);
+    } else {
+        VDP2CalcWindows<deinterlace, false>(y);
+    }
 
     // Draw sprite layer
-    (this->*fnDrawSprite[colorMode][rotate][false])(y);
+    (this->*fnDrawSprite[colorMode][rotate][altField])(y);
 
     // Draw background layers
     if (regs2.bgEnabled[5]) {
-        VDP2DrawRotationBG<0>(y, colorMode, false); // RBG0
-        VDP2DrawRotationBG<1>(y, colorMode, false); // RBG1
+        VDP2DrawRotationBG<0>(y, colorMode, altField); // RBG0
+        VDP2DrawRotationBG<1>(y, colorMode, altField); // RBG1
     } else {
-        // Update line screen scroll coordinates for NBG0 and NBG1
-        for (uint32 i = 0; i < 2; ++i) {
-            if (!m_layerEnabled[i + 2]) {
-                continue;
-            }
-            const BGParams &bgParams = regs2.bgParams[i + 1];
-            NormBGLayerState &bgState = m_normBGLayerStates[i];
-            VDP2UpdateLineScreenScroll(y, bgParams, bgState);
-        }
-
-        VDP2DrawRotationBG<0>(y, colorMode, false); // RBG0
+        VDP2DrawRotationBG<0>(y, colorMode, altField); // RBG0
         if (interlaced) {
-            VDP2DrawNormalBG<0, deinterlace>(y, colorMode, false); // NBG0
-            VDP2DrawNormalBG<1, deinterlace>(y, colorMode, false); // NBG1
-            VDP2DrawNormalBG<2, deinterlace>(y, colorMode, false); // NBG2
-            VDP2DrawNormalBG<3, deinterlace>(y, colorMode, false); // NBG3
+            VDP2DrawNormalBG<0, deinterlace>(y, colorMode, altField); // NBG0
+            VDP2DrawNormalBG<1, deinterlace>(y, colorMode, altField); // NBG1
+            VDP2DrawNormalBG<2, deinterlace>(y, colorMode, altField); // NBG2
+            VDP2DrawNormalBG<3, deinterlace>(y, colorMode, altField); // NBG3
         } else {
-            VDP2DrawNormalBG<0, false>(y, colorMode, false); // NBG0
-            VDP2DrawNormalBG<1, false>(y, colorMode, false); // NBG1
-            VDP2DrawNormalBG<2, false>(y, colorMode, false); // NBG2
-            VDP2DrawNormalBG<3, false>(y, colorMode, false); // NBG3
-        }
-
-        // Update NBG coordinates
-        for (uint32 i = 0; i < 4; ++i) {
-            if (!m_layerEnabled[i + 2]) {
-                continue;
-            }
-            const BGParams &bgParams = regs2.bgParams[i + 1];
-            NormBGLayerState &bgState = m_normBGLayerStates[i];
-            bgState.fracScrollY += bgParams.scrollIncV;
-            // Update the vertical scroll coordinate twice in double-density interlaced mode.
-            // If deinterlacing, the second increment is done after rendering the alternate scanline.
-            if (!deinterlace && doubleDensity) {
-                bgState.fracScrollY += bgParams.scrollIncV;
-            }
+            VDP2DrawNormalBG<0, false>(y, colorMode, altField); // NBG0
+            VDP2DrawNormalBG<1, false>(y, colorMode, altField); // NBG1
+            VDP2DrawNormalBG<2, false>(y, colorMode, altField); // NBG2
+            VDP2DrawNormalBG<3, false>(y, colorMode, altField); // NBG3
         }
     }
 
     // Compose image
-    VDP2ComposeLine<deinterlace, transparentMeshes>(y, false, false);
-
-    // Draw complementary field if deinterlace is enabled while in interlaced modes
-    if constexpr (deinterlace) {
-        if (interlaced) {
-            // The alternate VDP2 line only needs to be redrawn in double-density mode
-            if (doubleDensity) {
-                // Precalculate window state
-                VDP2CalcWindows<true, true>(y);
-            }
-
-            // Draw sprite layer
-            (this->*fnDrawSprite[colorMode][rotate][doubleDensity])(y);
-
-            if (doubleDensity) {
-                // Draw background layers
-                if (regs2.bgEnabled[5]) {
-                    VDP2DrawRotationBG<0>(y, colorMode, true); // RBG0
-                    VDP2DrawRotationBG<1>(y, colorMode, true); // RBG1
-                } else {
-                    VDP2DrawRotationBG<0>(y, colorMode, true);     // RBG0
-                    VDP2DrawNormalBG<0, true>(y, colorMode, true); // NBG0
-                    VDP2DrawNormalBG<1, true>(y, colorMode, true); // NBG1
-                    VDP2DrawNormalBG<2, true>(y, colorMode, true); // NBG2
-                    VDP2DrawNormalBG<3, true>(y, colorMode, true); // NBG3
-
-                    // Update NBG coordinates
-                    for (uint32 i = 0; i < 4; ++i) {
-                        if (!m_layerEnabled[i + 2]) {
-                            continue;
-                        }
-                        const BGParams &bgParams = regs2.bgParams[i + 1];
-                        NormBGLayerState &bgState = m_normBGLayerStates[i];
-                        bgState.fracScrollY += bgParams.scrollIncV;
-                    }
-                }
-            }
-
-            // Compose image
-            VDP2ComposeLine<true, transparentMeshes>(y, doubleDensity, true);
-        }
-    }
+    VDP2ComposeLine<deinterlace, transparentMeshes>(y, doubleDensity && altField, altField);
 }
 
 FORCE_INLINE void VDP::VDP2DrawLineColorAndBackScreens(uint32 y) {
@@ -3477,12 +3509,15 @@ template <uint32 bgIndex, bool deinterlace>
 FORCE_INLINE void VDP::VDP2DrawNormalBG(uint32 y, uint32 colorMode, bool altField) {
     static_assert(bgIndex < 4, "Invalid NBG index");
 
-    using FnDraw = void (VDP::*)(uint32 y, const BGParams &, LayerState &, NormBGLayerState &, std::span<const bool>);
+    using FnDrawScroll = void (VDP::*)(uint32 y, const BGParams &, LayerState &, const NormBGLayerState &,
+                                       CharacterFetcherState &, std::span<const bool>, bool);
+    using FnDrawBitmap =
+        void (VDP::*)(uint32 y, const BGParams &, LayerState &, const NormBGLayerState &, std::span<const bool>);
 
     // Lookup table of scroll BG drawing functions
     // Indexing: [charMode][fourCellChar][colorFormat][colorMode]
     static constexpr auto fnDrawScroll = [] {
-        std::array<std::array<std::array<std::array<FnDraw, 4>, 8>, 2>, 3> arr{};
+        std::array<std::array<std::array<std::array<FnDrawScroll, 4>, 8>, 2>, 3> arr{};
 
         util::constexpr_for<3 * 2 * 8 * 4>([&](auto index) {
             const uint32 fcc = bit::extract<0>(index());
@@ -3502,7 +3537,7 @@ FORCE_INLINE void VDP::VDP2DrawNormalBG(uint32 y, uint32 colorMode, bool altFiel
     // Lookup table of bitmap BG drawing functions
     // Indexing: [colorFormat]
     static constexpr auto fnDrawBitmap = [] {
-        std::array<std::array<FnDraw, 4>, 8> arr{};
+        std::array<std::array<FnDrawBitmap, 4>, 8> arr{};
 
         util::constexpr_for<8 * 4>([&](auto index) {
             const uint32 cf = bit::extract<0, 2>(index());
@@ -3523,7 +3558,8 @@ FORCE_INLINE void VDP::VDP2DrawNormalBG(uint32 y, uint32 colorMode, bool altFiel
     const VDP2Regs &regs = VDP2GetRegs();
     const BGParams &bgParams = regs.bgParams[bgIndex + 1];
     LayerState &layerState = m_layerStates[altField][bgIndex + 2];
-    NormBGLayerState &bgState = m_normBGLayerStates[bgIndex];
+    const NormBGLayerState &bgState = m_normBGLayerStates[bgIndex];
+    CharacterFetcherState &charState = m_charFetchers[altField][bgIndex];
     auto windowState = std::span<const bool>{m_bgWindows[altField][bgIndex + 1]}.first(m_HRes);
 
     const uint32 cf = static_cast<uint32>(bgParams.colorFormat);
@@ -3536,14 +3572,8 @@ FORCE_INLINE void VDP::VDP2DrawNormalBG(uint32 y, uint32 colorMode, bool altFiel
         const uint32 chm = static_cast<uint32>(twc   ? CharacterMode::TwoWord
                                                : exc ? CharacterMode::OneWordExtended
                                                      : CharacterMode::OneWordStandard);
-        (this->*fnDrawScroll[chm][fcc][cf][colorMode])(y, bgParams, layerState, bgState, windowState);
-    }
-
-    if (bgParams.mosaicEnable) {
-        ++bgState.mosaicCounterY;
-        if (bgState.mosaicCounterY >= regs.mosaicV) {
-            bgState.mosaicCounterY = 0;
-        }
+        (this->*fnDrawScroll[chm][fcc][cf][colorMode])(y, bgParams, layerState, bgState, charState, windowState,
+                                                       altField);
     }
 }
 
@@ -3553,12 +3583,14 @@ FORCE_INLINE void VDP::VDP2DrawRotationBG(uint32 y, uint32 colorMode, bool altFi
 
     static constexpr bool selRotParam = bgIndex == 0;
 
-    using FnDraw = void (VDP::*)(uint32 y, const BGParams &, LayerState &, std::span<const bool>, bool);
+    using FnDrawScroll =
+        void (VDP::*)(uint32 y, const BGParams &, LayerState &, CharacterFetcherState &, std::span<const bool>, bool);
+    using FnDrawBitmap = void (VDP::*)(uint32 y, const BGParams &, LayerState &, std::span<const bool>, bool);
 
     // Lookup table of scroll BG drawing functions
     // Indexing: [charMode][fourCellChar][colorFormat][colorMode]
     static constexpr auto fnDrawScroll = [] {
-        std::array<std::array<std::array<std::array<FnDraw, 4>, 8>, 2>, 3> arr{};
+        std::array<std::array<std::array<std::array<FnDrawScroll, 4>, 8>, 2>, 3> arr{};
 
         util::constexpr_for<3 * 2 * 8 * 4>([&](auto index) {
             const uint32 fcc = bit::extract<0>(index());
@@ -3578,7 +3610,7 @@ FORCE_INLINE void VDP::VDP2DrawRotationBG(uint32 y, uint32 colorMode, bool altFi
     // Lookup table of bitmap BG drawing functions
     // Indexing: [colorFormat][colorMode]
     static constexpr auto fnDrawBitmap = [] {
-        std::array<std::array<FnDraw, 4>, 8> arr{};
+        std::array<std::array<FnDrawBitmap, 4>, 8> arr{};
 
         util::constexpr_for<8 * 4>([&](auto index) {
             const uint32 cf = bit::extract<0, 2>(index());
@@ -3599,6 +3631,7 @@ FORCE_INLINE void VDP::VDP2DrawRotationBG(uint32 y, uint32 colorMode, bool altFi
     const VDP2Regs &regs = VDP2GetRegs();
     const BGParams &bgParams = regs.bgParams[bgIndex];
     LayerState &layerState = m_layerStates[altField][bgIndex + 1];
+    CharacterFetcherState &charState = m_charFetchers[altField][bgIndex + 4];
     auto windowState = std::span<const bool>{m_bgWindows[altField][bgIndex]}.first(m_HRes);
 
     const uint32 cf = static_cast<uint32>(bgParams.colorFormat);
@@ -3611,7 +3644,7 @@ FORCE_INLINE void VDP::VDP2DrawRotationBG(uint32 y, uint32 colorMode, bool altFi
         const uint32 chm = static_cast<uint32>(twc   ? CharacterMode::TwoWord
                                                : exc ? CharacterMode::OneWordExtended
                                                      : CharacterMode::OneWordStandard);
-        (this->*fnDrawScroll[chm][fcc][cf][colorMode])(y, bgParams, layerState, windowState, altField);
+        (this->*fnDrawScroll[chm][fcc][cf][colorMode])(y, bgParams, layerState, charState, windowState, altField);
     }
 }
 
@@ -4776,11 +4809,13 @@ FORCE_INLINE void VDP::VDP2ComposeLine(uint32 y, bool altFieldSrc, bool altField
 
 template <VDP::CharacterMode charMode, bool fourCellChar, ColorFormat colorFormat, uint32 colorMode, bool deinterlace>
 NO_INLINE void VDP::VDP2DrawNormalScrollBG(uint32 y, const BGParams &bgParams, LayerState &layerState,
-                                           NormBGLayerState &bgState, std::span<const bool> windowState) {
+                                           const NormBGLayerState &bgState, CharacterFetcherState &charState,
+                                           std::span<const bool> windowState, bool altField) {
     const VDP2Regs &regs = VDP2GetRegs();
 
+    const bool altLine = deinterlace && altField && regs.TVMD.LSMDn == InterlaceMode::DoubleDensity;
     uint32 fracScrollX = bgState.fracScrollX + bgParams.scrollAmountH;
-    const uint32 fracScrollY = bgState.fracScrollY + bgState.scrollAmountV;
+    const uint32 fracScrollY = bgState.fracScrollY + bgState.scrollAmountV + (altLine ? bgParams.scrollIncV : 0);
 
     uint32 cellScrollTableAddress = regs.verticalCellScrollTableAddress + bgState.vertCellScrollOffset;
 
@@ -4835,8 +4870,7 @@ NO_INLINE void VDP::VDP2DrawNormalScrollBG(uint32 y, const BGParams &bgParams, L
 
             // Plot pixel
             const Pixel pixel = VDP2FetchScrollBGPixel<false, charMode, fourCellChar, colorFormat, colorMode>(
-                bgParams, bgParams.pageBaseAddresses, bgParams.pageShiftH, bgParams.pageShiftV, scrollCoord,
-                bgState.charFetcher);
+                bgParams, bgParams.pageBaseAddresses, bgParams.pageShiftH, bgParams.pageShiftV, scrollCoord, charState);
             layerState.pixels.SetPixel(x, pixel);
         }
 
@@ -4863,8 +4897,7 @@ NO_INLINE void VDP::VDP2DrawNormalScrollBG(uint32 y, const BGParams &bgParams, L
 
         // Fetch pixel
         VDP2FetchScrollBGPixel<false, charMode, fourCellChar, colorFormat, colorMode>(
-            bgParams, bgParams.pageBaseAddresses, bgParams.pageShiftH, bgParams.pageShiftV, scrollCoord,
-            bgState.charFetcher);
+            bgParams, bgParams.pageBaseAddresses, bgParams.pageShiftH, bgParams.pageShiftV, scrollCoord, charState);
 
         // Increment horizontal coordinate
         fracScrollX += bgState.scrollIncH * 8;
@@ -4873,7 +4906,7 @@ NO_INLINE void VDP::VDP2DrawNormalScrollBG(uint32 y, const BGParams &bgParams, L
 
 template <ColorFormat colorFormat, uint32 colorMode, bool deinterlace>
 NO_INLINE void VDP::VDP2DrawNormalBitmapBG(uint32 y, const BGParams &bgParams, LayerState &layerState,
-                                           NormBGLayerState &bgState, std::span<const bool> windowState) {
+                                           const NormBGLayerState &bgState, std::span<const bool> windowState) {
     const VDP2Regs &regs = VDP2GetRegs();
 
     uint32 fracScrollX = bgState.fracScrollX + bgParams.scrollAmountH;
@@ -4943,7 +4976,8 @@ NO_INLINE void VDP::VDP2DrawNormalBitmapBG(uint32 y, const BGParams &bgParams, L
 
 template <bool selRotParam, VDP::CharacterMode charMode, bool fourCellChar, ColorFormat colorFormat, uint32 colorMode>
 NO_INLINE void VDP::VDP2DrawRotationScrollBG(uint32 y, const BGParams &bgParams, LayerState &layerState,
-                                             std::span<const bool> windowState, bool altField) {
+                                             CharacterFetcherState &charState, std::span<const bool> windowState,
+                                             bool altField) {
     const VDP2Regs &regs = VDP2GetRegs();
 
     const bool doubleResH = regs.TVMD.HRESOn & 0b010;
@@ -5011,7 +5045,7 @@ NO_INLINE void VDP::VDP2DrawRotationScrollBG(uint32 y, const BGParams &bgParams,
             // Plot pixel
             const Pixel pixel = VDP2FetchScrollBGPixel<true, charMode, fourCellChar, colorFormat, colorMode>(
                 bgParams, rotParamState.pageBaseAddresses, rotParams.pageShiftH, rotParams.pageShiftV, scrollCoord,
-                m_rotParamStates[rotParamSelector].charFetcher);
+                m_charFetchers[altField][rotParamSelector]);
             layerState.pixels.SetPixel(xx, pixel);
             if (doubleResH) {
                 layerState.pixels.SetPixel(xx + 1, pixel);